# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UniPath MIS - Genomics & Bioinformatics Project Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scientists:
    get:
      operationId: listScientists
      tags: [scientists]
      summary: List all scientists
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scientist"
    post:
      operationId: createScientist
      tags: [scientists]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScientistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scientist"
  /scientists/{id}:
    put:
      operationId: updateScientist
      tags: [scientists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScientistInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scientist"
    delete:
      operationId: deleteScientist
      tags: [scientists]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────────────────────────────────────────
  # TERRITORIES
  # ──────────────────────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects with filters and pagination
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
        - name: serviceId
          in: query
          schema:
            type: integer
        - name: scientistId
          in: query
          schema:
            type: integer
        - name: territoryId
          in: query
          schema:
            type: integer
        - name: salesPersonId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: month
          in: query
          schema:
            type: string
        - name: withAnalysis
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get full project details with QC, delivery, invoices, payments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────────────────────────────────────────
  # QC RECORDS
  # ──────────────────────────────────────────────────────────────────────────
  /projects/{projectId}/qc:
    get:
      operationId: getQcRecord
      tags: [qc]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QC record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcRecord"
    put:
      operationId: upsertQcRecord
      tags: [qc]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcRecordInput"
      responses:
        "200":
          description: Created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcRecord"

  # ──────────────────────────────────────────────────────────────────────────
  # DATA DELIVERY
  # ──────────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Scientist:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        designation:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    ScientistInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        designation:
          type: string
        email:
          type: string
...
    Project:
      type: object
      required:
        - id
        - projectCode
        - date
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectCode:
          type: string
        date:
          type: string
          format: date
        month:
          type: string
        labSubmissionDate:
          type: string
          format: date
        scientistId:
          type: integer
        scientistName:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        billingClientId:
          type: integer
        billingClientName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        serviceHead:
          type: string
        sampleType:
          type: string
        withAnalysis:
          type: string
        noOfSamples:
          type: integer
        dataRequirement:
          type: string
        gbPerSample:
          type: number
        totalGb:
          type: number
        ratePerSample:
          type: number
        totalAmount:
          type: number
        gst:
          type: number
        totalProjectCost:
          type: number
        quotationNo:
          type: string
        salesPersonId:
          type: integer
        salesPersonName:
          type: string
        territoryId:
          type: integer
        territoryName:
          type: string
        city:
          type: string
        status:
          type: string
        remark:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required:
        - projectCode
        - date
      properties:
        projectCode:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        month:
          type: string
        labSubmissionDate:
          type: string
          format: date
        scientistId:
          type: integer
        clientId:
          type: integer
        billingClientId:
          type: integer
        serviceId:
          type: integer
        sampleType:
          type: string
        withAnalysis:
          type: string
        noOfSamples:
          type: integer
        dataRequirement:
          type: string
        gbPerSample:
          type: number
        totalGb:
          type: number
        ratePerSample:
          type: number
        totalAmount:
          type: number
        gst:
          type: number
        totalProjectCost:
          type: number
        quotationNo:
          type: string
        salesPersonId:
          type: integer
        territoryId:
          type: integer
        city:
          type: string
        status:
          type: string
        remark:
          type: string
    ProjectUpdate:
      type: object
      properties:
        projectCode:
          type: string
        date:
          type: string
          format: date
        month:
          type: string
        labSubmissionDate:
          type: string
          format: date
        scientistId:
          type: integer
        clientId:
          type: integer
        billingClientId:
          type: integer
        serviceId:
          type: integer
        sampleType:
          type: string
        withAnalysis:
          type: string
        noOfSamples:
          type: integer
        dataRequirement:
          type: string
        gbPerSample:
          type: number
        totalGb:
          type: number
        ratePerSample:
          type: number
        totalAmount:
          type: number
        gst:
          type: number
        totalProjectCost:
          type: number
        quotationNo:
          type: string
        salesPersonId:
          type: integer
        territoryId:
          type: integer
        city:
          type: string
        status:
          type: string
        remark:
          type: string
    ProjectListResponse:
      type: object
      required:
        - data
        - total
        - page
        - pageSize
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    ProjectDetail:
      type: object
      required:
        - id
        - projectCode
        - date
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectCode:
          type: string
        date:
          type: string
          format: date
        month:
          type: string
        labSubmissionDate:
          type: string
          format: date
        scientistId:
          type: integer
        scientistName:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        billingClientId:
          type: integer
        billingClientName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        serviceHead:
          type: string
        sampleType:
          type: string
        withAnalysis:
          type: string
        noOfSamples:
          type: integer
        dataRequirement:
          type: string
        gbPerSample:
          type: number
        totalGb:
          type: number
        ratePerSample:
          type: number
        totalAmount:
          type: number
        gst:
          type: number
        totalProjectCost:
          type: number
        quotationNo:
          type: string
        salesPersonId:
          type: integer
        salesPersonName:
          type: string
        territoryId:
          type: integer
        territoryName:
          type: string
        city:
          type: string
        status:
          type: string
        remark:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        qcRecord:
          $ref: "#/components/schemas/QcRecord"
        dataDelivery:
          $ref: "#/components/schemas/DataDelivery"
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    QcRecord:
      type: object
      required:
        - id
        - projectId
      properties:
        id:
          type: integer
        projectId:
          type: integer
        qcPass:
          type: integer
        qcFail:
          type: integer
        qcReportDate:
          type: string
          format: date
        qcTatDays:
          type: integer
        qcTatStatus:
          type: string
        runNo:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QcRecordInput:
      type: object
      properties:
        qcPass:
          type: integer
        qcFail:
          type: integer
        qcReportDate:
          type: string
          format: date
        qcTatDays:
          type: integer
        qcTatStatus:
          type: string
        runNo:
          type: string
    DataDelivery:
      type: object
      required:
        - id
        - projectId
      properties:
        id:
          type: integer
        projectId:
          type: integer
        rawDataSentDate:
          type: string
          format: date
        finalDataDate:
          type: string
          format: date
        rawDataDays:
          type: integer
        finalDataDays:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Invoice:
      type: object
      required:
        - id
        - projectId
        - paymentStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectCode:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        invoiceNo:
          type: string
        invoiceDate:
          type: string
          format: date
        qcPassSamples:
          type: integer
        subtotal:
          type: number
        gst:
          type: number
        totalAmount:
          type: number
        invoiceTatDays:
          type: integer
        paymentStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Payment:
      type: object
      required:
        - id
        - projectId
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectId:
          type: integer
        invoiceId:
          type: integer
        invoiceNo:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        receivedAmount:
          type: number
        remainingAmount:
          type: number
        paymentReceivedDate:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...